Turn on forkMode=always for tests
Default is forkMode=once, meaning all test suites run in a separate process
from maven itself, but in the same process as all other test suites in the
same module. Because the call to SelectorProvider.provider() caches a result,
and because test suites can run in any order, this means that the
DelegateSelectorProvider would need to be initialized for each test suite that
runs in the same process. forkMode=always works around this.