Move max frame size off of NiftyClient and into connectSync/connectAsync parameters
This makes it easier to integrate into the swift client configuration settings, which are per client type. The expected max size for a frame will be different depending on the type of service.

Also changing default max frame size to 16 MB instead of 1 MB, as several of our services return more than 1 MB on a regular basis.