Remove ThriftClientManagerProvider
Creating the ThriftClientManager through a provider makes it harder for Guice life cycle managemers to close() the manager in a @PreDestroy method. The only reason it was there was to enable passing a non-default maxFrameSize.

This change makes maxFrameSize a per-client configuration parameter, and binds ThriftClientManager directly (not through a provider).