Separate swift annotations into their own artifact
This should help Finagle integration, minimizing compile-time dependnecies by not having the artifact that contains annotations depend on any of the other swift modules (which themselves depend on a bunch of other artifacts).

Going forward, this also should allows jars of only swift interfaces and types to be mostly version-agnostic as swift version moves (except when a new swift version changes existing annotations).