Add support for annotations on service declarations
FB C++ thrift added a priority annotation, which influences execution order of requests. In addition to annotating methods with priorities, it supports annotating services with priorities.

This change doesn't add support for priorities, just adds support for reading annotations attached to a service so that .thrift files using this syntax will still parse in the swift generator.