Add Union support to swift.
An union bean must be annotated with @ThriftUnion("name"), similar to @ThriftStruct.

In addtion, it must have a getter or a field marked with @ThriftUnionId. This is used when
serializing the object to determine which field is read. Fields are marked with @ThriftField just as
a @ThriftStruct, however only one field is ever written or read.