[swift2thrift] Support classes in different packages with a command-line flag
This is useful if you want to generate one thrift file from java classes
in different packges.

Usually it's better to generate one thrift file per package, but if one
thrift file is easier to deal with, generating swift files back from it
is not necessary, and the thrift file will only be used to generate code
in other languages where namespacing is not an issue, add an option to
have one thrift file.