Make it an error to have multiple extracted or multiple specified names for a single field
The existing code was trying to allow for the cases that some appearances of the field might have been explicitly renamed in a @ThriftField annotation, while others may not have been, so there could be conflicts between the set of names retreived by getOrExtractThriftFieldName(). For this reason, conflicting names was only a warning.

However it should be an error to have multiple explicit names, or multiple implicit names. This change checks each separately to make sure everything is consistent.