Revert "Make it an error to have multiple extracted or multiple specified names for a single field"
This reverts commit fec043189859fef6628dabfa12c688e62a1ad58e.

There are problems with this change in that I thought "extractName()" always gets the implicit name (not the name in the annotation) but this is not true for ParameterInjections, where the annotation name takes precedence. This is fixable but I'm not sure it's worth it at this time, so I'll revert this diff, making it a warning again to have multiple names for a thrift field.