[codec] Allow negative field numbers in structs/unions
Summary:

Negative field IDs are strongly discouraged, but some old services used
them and are stuck with them to maintain wire compatibility with older
clients. Don't error out on them. Add an `isLegacyId` attribute to the
`@ThriftField` annotation to allow them when necessary. Any
`@ThriftField` which has a negative ID must have `isLegacyId=true`;
`@ThriftField`'s with non-negative IDs must not (but may specify
`isLegacyId=false` if they want.)

Test Plan:

grep in swift-codec for `> 0`, `>= 0`, `< 0`, `<= 0`, (and the other way
around). A bunch of ThriftStructs which didn't used to be able to be
loaded are now. Added unit tests.

Reviewers: alandau, andrewcox, haijunz

CC: