[codec] remove the 'structs must be final' check
Summary:

1. The requirement is unnecessary. Swift operates fine without it
(except in one circumstance with exceptions, fixed in the followup
diff.)

2. It breaks some valid usage patterns, like subclassing ThriftStructs
with implementation-internal fields, or using ThriftStructs as mocks or
spies in mockito.

3. It's unconventional, from both a Java perspective and from the
perspective of most (if not all) other implementations of Thrift in OO
languages.

4. The original motivation for adding the requirement was because
swift-generator wanted to add equals and hashCode, and final is required
(in some sense) to make those make sense. However, putting the
restriction in swift-codec is unnecessary; it's sufficient for
swift-generator to declare the structs it generates as final (and it
does.)

Test Plan: Unit tests included. Also, works in a larger project.

Reviewers: alandau, andrewcox, haijunz, dphillips

CC: