[service] support throwing subclasses of exceptions
Summary: Previously, if an implementation of a ThriftMethod threw an
exception, we looked up the codec by checking if its class is the same
as any of the declared types. That doesn't work if the method throws a
subclass of the declared type. Now, if the first lookup fails, see if
the exception is a subclass of any of the declared exceptions.

Test Plan: Unit tests included

Reviewers: alandau, andrewcox, haijunz

CC: