Fix swift-maven-plugin integration tests
This test downloads alternate versions of maven to test the plugin against, from a maven repo. Since it was running inside surefire (which for uniform testing purposes tries to provdie a clean environment separate from the user's environment) it was not respecting network properties configured by the user and so the download could fail if you were on an IPv6-only machine, or behind a proxy, etc.