Add SessionAwareSslHandler
This adds an SessionAwareSslHandler. Many components
might want to get access to an ssl session. The sslhandler
API is currently a bit racy, because calling it twice could
potentially cause renegotiation to trigger.

This changes it so that upstream filters will not get a
channelConnected until the SSL connection is established.

SessionAwareSslHandler propagates a message called sslsession
which upstream handlers can also register for.