Fix generated Java code for Thrift unions with multiple fields of same type
Currently, when generating Java code for Thrift unions with multiple fields
of the same type, the code won't compile. This is because Swift generates
a constructor per field, so multiple constructors will have the same
arguments even though they initialize fields differently.

This patch changes the union template to use a builder paradigm in order
to allow for this case.