Fix TestSwiftGenerator
TestSwiftGenerator is a test application that generates swift classes for several test IDLs.
Those IDLs are located in resources section of the project and there was a previous commit
that changed location of those test IDL files.
After that change swift generator could not find included IDLs which is fixed with this change.