Allow fallback to plain java namespace if java.swift is missing
There projects that are not using swift directly so they do not have java.swift namespace in their IDLs.
And this causes problems for projects that are using swift and including such IDLs indirectly.
One solution is to modify IDL file and add java.swift namespace which is not always possible or convenient.
Another approach is to use plain java namespace and generate swift classes using it.
This change introduces later option adding new boolean tweak configuration option.
For command line tool it is `fallback_to_java_namespace` and `fallbackToPlainJavaNamespace` for maven plugin.